Параметризованный тест выделен в отдельный класс BurgerGetPriceParameterizedTest, исправлен код в BurgerTest, изменен pom.xml
