+++ pom.xml
@@ -67,5 +67,12 @@
             <version>5.19.0</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.assertj</groupId>
+            <artifactId>assertj-core</artifactId>
+            <version>3.27.6</version>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 </project>
